Refactor hotpatch_vars and apply it to libtorch
Remove new empty line
